refactor: Separate handlers and the main app
